Use latest Cloud Vision sample with Label, Landmark, and Face detection
