+++ vision/Gemfile
@@ -1,9 +1,25 @@
+# Copyright 2016 Google, Inc
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 source "https://rubygems.org"
 
-gem 'google-cloud', '~> 0.20.0'
+gem "google-cloud-vision"
+gem "rmagick"
 
 group :test do
   gem "rake"
   gem "rubocop"
   gem "rspec"
+  gem "rack-test"
 end
