+++ vision/detect.rb
